requirements.txt
aiohttp==3.8.5
bs4==0.0.1
lxml==4.9.3
peewee==3.16.3
psycopg2-binary==2.9.7
python-dotenv==1.0.0